Dashboard: patience_counter is f64 — magnitude-aware patience broke the iterations endpoint
Magnitude-aware patience (11/07) writes FRACTIONAL counters (4.618/5) as REAL;
the i32-only row decode failed on any such row and the whole
/api/experiments/{id}/iterations response stream-reset -> UI showed 'No
iterations recorded' despite 90 rows in the DB (found on experiment 10185).
Decode now tries f64 then i64 (old INTEGER rows still fine); frontend type was
already number|null. Deployed 12/07 (dashboard restart; worker + running flow
4416 unaffected — P3 stale-recovery semantics held).

+++ dashboard/src/models/mod.rs
@@ -317,7 +317,11 @@ pub struct Iteration {
     pub baseline_ce: Option<f64>,
     pub delta_baseline: Option<f64>,
     pub delta_previous: Option<f64>,
-    pub patience_counter: Option<i32>,
+    // f64 since 12/07/2026: magnitude-aware patience records FRACTIONAL
+    // counters (e.g. 4.618/5); older rows are INTEGER — read via
+    // try_get::<f64> fallback in db/iterations.rs (sqlite REAL/INTEGER both
+    // decode as f64 there).
+    pub patience_counter: Option<f64>,
     pub patience_max: Option<i32>,
     pub candidates_total: Option<i32>,
     pub created_at: DateTime<Utc>,
